+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.9.5
 
+* more pydoc for `satella.files`
+* add `try_unlink`